chore(net,vpcmap): Move VpcDiscrimant to net
Precursor to using VpcDiscriminant instead of exclusively VNI
